+++ packages/blocks/core/js/bootstrap.js
@@ -1,6 +1,7 @@
 import {
 	bootstrapOutermostIconBlock,
 	bootstrapBlocksyBreadcrumbs,
+	bootstrapBlocksyAboutMe,
 } from './third-party/bootstraps';
 import { bootstrapSocialLinksCoreBlock } from './wordpress/bootstraps';
 
@@ -8,4 +9,5 @@ export function blockeraBootstrapBlocks() {
 	bootstrapSocialLinksCoreBlock();
 	bootstrapOutermostIconBlock();
 	bootstrapBlocksyBreadcrumbs();
+	bootstrapBlocksyAboutMe();
 }