+++ packages/blocks/core/js/third-party/blocksy-about-me/index.js
@@ -1,0 +1,39 @@
+// @flow
+
+/**
+ * External dependencies
+ */
+import { __ } from '@wordpress/i18n';
+
+/**
+ * Blockera dependencies
+ */
+import { SharedBlockExtension } from '@blockera/editor';
+import { Icon } from '@blockera/icons';
+
+/**
+ * Internal dependencies
+ */
+import type { BlockType } from '../../type';
+
+export const BlocksyAboutMe: BlockType = {
+	name: 'blockeraBlocksyAboutMe',
+	targetBlock: 'blocksy/about-me',
+	blockeraInnerBlocks: {
+		'elements/text': {
+			name: 'elements/text',
+			label: __('Texts', 'blockera'),
+			description: __(
+				'All text elements inside the about me block.',
+				'blockera'
+			),
+			icon: <Icon icon="block-blocksy-about-texts" iconSize="20" />,
+			settings: {
+				force: true,
+			},
+		},
+	},
+	edit: (props) => {
+		return <SharedBlockExtension {...props} />;
+	},
+};