+++ packages/blocks/core/php/third-party/blocksy-about-me/block.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * Configure block all arguments.
+ *
+ * @var array $args the block arguments!
+ *
+ * @package blockera/packages/blocks/php/third-party/blocksy-about-me
+ */
+
+return array_merge(
+	$args,
+	[
+		'selectors'  => array_merge(
+			$args['selectors'] ?? [],
+			[
+				'blockera/elements/text'   => [
+					'root' => '.ct-about-me-name span, .ct-about-me-name a',
+				],
+			]
+		),
+	]
+);